Row search
Added the basic functionality for code. Can check for words that fit in rows.

New functions added. Rowchecker to find words that fits in rows.

Rowwriter to rewrite the crossword pattern with words.

Arrayprint to print the crossword that is modified.

Still have to add functions to find words for columns.